code/lambdas/fragmentation_producer/requirements.txt
boto3
kafka-python
pandas
faker
aws-msk-iam-sasl-signer-python